requirements.txt
fastapi==0.95.0
uvicorn==0.22.0
sqlalchemy==1.4.46
pillow==9.5.0
numpy==1.25.2
torch==2.0.1
torchvision==0.15.2
facenet-pytorch==2.5.2